back-end/requirements.txt
Flask==1.1.2
flask-restful
SQLAlchemy
requests
flask-cors
flask-SQLAlchemy
flask-marshmallow
flask-restless
marshmallow-sqlalchemy
gunicorn
black
selenium
python-dotenv
uWSGI==2.0.19.1